# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AI Customer Support Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResult"
        "401":
          description: Invalid credentials
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user and organization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Registration successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResult"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current authenticated user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /auth/refresh:
    post:
      operationId: refreshToken
      tags: [auth]
      summary: Refresh access token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshInput"
      responses:
        "200":
          description: New tokens
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResult"

  # ─── ORGANIZATIONS ──────────────────────────────────────────────────────────
...
  /conversations:
    get:
      operationId: listConversations
      tags: [conversations]
      summary: List conversations
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [open, resolved, pending, snoozed]
        - name: channelId
          in: query
          schema:
            type: number
        - name: contactId
          in: query
          schema:
            type: number
        - name: assigneeId
          in: query
          schema:
            type: number
      responses:
        "200":
          description: List of conversations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Conversation"
    post:
      operationId: createConversation
      tags: [conversations]
      summary: Create a conversation
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConversationInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Conversation"
...
  /channels:
    get:
      operationId: listChannels
      tags: [channels]
      summary: List channels
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of channels
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Channel"
    post:
      operationId: createChannel
      tags: [channels]
      summary: Create a channel
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChannelInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Channel"
  /channels/{id}:
    get:
      operationId: getChannel
      tags: [channels]
      summary: Get channel by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Channel
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Channel"
    patch:
      operationId: updateChannel
      tags: [channels]
      summary: Update a channel
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChannelUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Channel"
    delete:
      operationId: deleteChannel
      tags: [channels]
      summary: Delete a channel
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "204":
          description: Deleted

  # ─── NOTES ──────────────────────────────────────────────────────────────────
...
  /ai-settings:
    get:
      operationId: getAiSettings
      tags: [ai-settings]
      summary: Get AI settings for organization
      security:
        - bearerAuth: []
      responses:
        "200":
          description: AI settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiSettings"
    put:
      operationId: updateAiSettings
      tags: [ai-settings]
      summary: Update AI settings
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AiSettingsInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiSettings"

  # ─── ANALYTICS ──────────────────────────────────────────────────────────────
...
  /analytics/conversations:
    get:
      operationId: getConversationAnalytics
      tags: [analytics]
      summary: Get conversation analytics over time
      security:
        - bearerAuth: []
      parameters:
        - name: period
          in: query
          schema:
            type: string
            enum: [7d, 30d, 90d]
            default: "30d"
      responses:
        "200":
          description: Analytics data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConversationAnalytics"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string

        password:
          type: string
    RegisterInput:
      type: object
      required: [email, password, firstName, lastName, organizationName]
      properties:
        email:
          type: string

        password:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        organizationName:
          type: string
    RefreshInput:
      type: object
      required: [refreshToken]
      properties:
        refreshToken:
          type: string
    AuthResult:
      type: object
      required: [accessToken, refreshToken, user]
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        user:
          $ref: "#/components/schemas/User"

    # ─── ORGANIZATION ─────────────────────────────────────────────────────────
...
    User:
      type: object
      required: [id, email, firstName, lastName, role, isActive, createdAt]
      properties:
        id:
          type: number
        email:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        avatarUrl:
          type: ["string", "null"]
        role:
          type: string
          enum: [owner, admin, agent]
        isActive:
          type: boolean
        organizationId:
          type: ["number", "null"]
        createdAt:
          type: string
          format: date-time
...
    Contact:
      type: object
      required: [id, firstName, lastName, createdAt]
      properties:
        id:
          type: number
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        companyId:
          type: ["number", "null"]
        companyName:
          type: ["string", "null"]
        tags:
          type: array
          items:
            $ref: "#/components/schemas/Tag"
        lastActivityAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
...
    Conversation:
      type: object
      required: [id, status, channelType, createdAt]
      properties:
        id:
          type: number
        status:
          type: string
          enum: [open, resolved, pending, snoozed]
        channelType:
          type: string
          enum: [whatsapp, messenger, instagram, telegram, tiktok, web]
        channelId:
          type: ["number", "null"]
        contactId:
          type: ["number", "null"]
        contact:
          $ref: "#/components/schemas/Contact"
        assigneeId:
          type: ["number", "null"]
        assignee:
          $ref: "#/components/schemas/User"
        subject:
          type: ["string", "null"]
        lastMessageAt:
          type: ["string", "null"]
          format: date-time
        lastMessage:
          type: ["string", "null"]
        unreadCount:
          type: number
        aiHandled:
          type: boolean
        createdAt:
          type: string
          format: date-time
    ConversationInput:
      type: object
      required: [channelType]
      properties:
        channelType:
          type: string
          enum: [whatsapp, messenger, instagram, telegram, tiktok, web]
        channelId:
          type: number
        contactId:
          type: number
        subject:
          type: string
        assigneeId:
          type: number
...
    Channel:
      type: object
      required: [id, name, channelType, provider, isActive, createdAt]
      properties:
        id:
          type: number
        name:
          type: string
        channelType:
          type: string
          enum: [whatsapp, messenger, instagram, telegram, tiktok, web]
        provider:
          type: string
          enum: [whatsapp_cloud, whatsapp_web, meta_graph, instagram_graph, telegram_bot, tiktok_dm, web_widget]
        isActive:
          type: boolean
        phoneNumber:
          type: ["string", "null"]
        pageId:
          type: ["string", "null"]
        conversationCount:
          type: number
        createdAt:
          type: string
          format: date-time
    ChannelInput:
      type: object
      required: [name, channelType, provider]
      properties:
        name:
          type: string
        channelType:
          type: string
          enum: [whatsapp, messenger, instagram, telegram, tiktok, web]
        provider:
          type: string
          enum: [whatsapp_cloud, whatsapp_web, meta_graph, instagram_graph, telegram_bot, tiktok_dm, web_widget]
    ChannelUpdate:
      type: object
      properties:
        name:
          type: string
        isActive:
          type: boolean

    # ─── NOTE ─────────────────────────────────────────────────────────────────
...
    Tag:
      type: object
      required: [id, name, color]
      properties:
        id:
          type: number
        name:
          type: string
        color:
          type: string
...
    AiSettings:
      type: object
      required: [id, organizationId, provider, model, autoReplyEnabled]
      properties:
        id:
          type: number
        organizationId:
          type: number
        provider:
          type: string
          enum: [ollama, openai_compatible]
        model:
          type: string
        baseUrl:
          type: ["string", "null"]
        systemPrompt:
          type: ["string", "null"]
        autoReplyEnabled:
          type: boolean
        confidenceThreshold:
          type: number
        escalateOnLowConfidence:
          type: boolean
        contextWindowSize:
          type: number
        maxTokens:
          type: number
        temperature:
          type: number
    AiSettingsInput:
      type: object
      properties:
        provider:
          type: string
          enum: [ollama, openai_compatible]
        model:
          type: string
        baseUrl:
          type: string
        systemPrompt:
          type: string
        autoReplyEnabled:
          type: boolean
        confidenceThreshold:
          type: number
        escalateOnLowConfidence:
          type: boolean
        contextWindowSize:
          type: number
        maxTokens:
          type: number
        temperature:
          type: number

    # ─── ANALYTICS ────────────────────────────────────────────────────────────
...
    ConversationAnalytics:
      type: object
      required: [dataPoints, summary]
      properties:
        dataPoints:
          type: array
          items:
            $ref: "#/components/schemas/AnalyticsDataPoint"
        summary:
          type: object
          required: [total, resolved, avgResponseTime, aiHandled]
          properties:
            total:
              type: number
            resolved:
              type: number
            avgResponseTime:
              type: number
            aiHandled:
              type: number
    AnalyticsDataPoint:
      type: object
      required: [date, opened, resolved, aiHandled]
      properties:
        date:
          type: string

        opened:
          type: number
        resolved:
          type: number
        aiHandled:
          type: number
...