# lib/api-zod/src/generated/types/analyticsDataPoint.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * AI Customer Support Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface AnalyticsDataPoint {
  date: string;
  opened: number;
  resolved: number;
  aiHandled: number;
}