# lib/api-zod/src/generated/types/authResult.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * AI Customer Support Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { User } from './user';

export interface AuthResult {
  accessToken: string;
  refreshToken: string;
  user: User;
}